+++ util.h
@@ -8,22 +8,22 @@
 #define EPSILON 0.00001
 #define PI 3.14159265358979323846f
 
-float absolute(float f){
+double absolute(double f){
     if(f < 0){
         f = f * -1;
     }
     return f;
 }
 
-bool equal(float f1, float f2){
+bool equal(double f1, double f2){
     return absolute(f1 - f2) < EPSILON;
 }
 
-float radians(float degrees){
+double radians(double degrees){
     return (degrees / 180) * PI;
 }
 
-float clamp(float val, float min, float max){
+double clamp(double val, double min, double max){
     if(val > max){
         return max;
     }
